WellcomeHome/Bot/requirements.txt
aiohttp
botbuilder-core
botbuilder-integration-aiohttp
botbuilder-dialogs
requests_oauthlib
azure-identity
azure-keyvault-secrets
azure-keyvault-keys
azure-cosmos
azure-storage-blob